Move plugins to separate files
