+++ packages/vite-plugin-cloudflare/src/context.ts
@@ -1,0 +1,7 @@
+import type { WorkerConfig } from "./plugin-config";
+import type { NodeJsCompat } from "./plugins/nodejs-compat";
+
+export type Context = Readonly<{
+	getWorkerConfig(environmentName: string): WorkerConfig | undefined;
+	getNodeJsCompat(environmentName: string): NodeJsCompat | undefined;
+}>;
