+++ packages/vite-plugin-cloudflare/src/miniflare-options.ts
@@ -28,7 +28,7 @@ import {
 	ROUTER_WORKER_NAME,
 	VITE_PROXY_WORKER_NAME,
 } from "./constants";
-import { additionalModuleRE } from "./shared";
+import { additionalModuleRE } from "./plugins/additional-modules";
 import { withTrailingSlash } from "./utils";
 import type { CloudflareDevEnvironment } from "./cloudflare-environment";
 import type {
