+++ packages/vite-plugin-cloudflare/src/shared.ts
@@ -1,14 +1,7 @@
-import { ADDITIONAL_MODULE_TYPES, virtualPrefix } from "./constants";
+import { virtualPrefix } from "./constants";
 
 export const UNKNOWN_HOST = "http://localhost";
 export const INIT_PATH = "/__vite_plugin_cloudflare_init__";
-
-const ADDITIONAL_MODULE_PATTERN = `__CLOUDFLARE_MODULE__(${ADDITIONAL_MODULE_TYPES.join("|")})__(.*?)__CLOUDFLARE_MODULE__`;
-export const additionalModuleRE = new RegExp(ADDITIONAL_MODULE_PATTERN);
-export const additionalModuleGlobalRE = new RegExp(
-	ADDITIONAL_MODULE_PATTERN,
-	"g"
-);
 
 // headers
 export const WORKER_ENTRY_PATH_HEADER = "__VITE_WORKER_ENTRY_PATH__";
